% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clean_R_package_data}
\alias{clean_R_package_data}
\title{Clean Data for the R Package}
\format{
A row of multiple heights in cm.
}
\usage{
data(clean_R_package_data)
}
\description{
A cleaned version of the raw dataset.
}
\keyword{datasets}
