% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_normalization.R
\name{data_normalization}
\alias{data_normalization}
\title{Data Normalization}
\usage{
data_normalization(dataset)
}
\arguments{
\item{x}{A vector.}
}
\value{
A normalized vector.
}
\description{
This function normalizes a vector to have a mean of 0 and standard deviation of 1.
}
\examples{
data_normalization(c(165,175,185))
}
