% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_summarisation.R
\name{data_summarise}
\alias{data_summarise}
\title{Data summarised}
\usage{
data_summarise(dataset)
}
\arguments{
\item{x}{A vector.}
}
\value{
A list with summary statistics.
}
\description{
This function provides summary statistics for a vector.
}
\examples{
data_summarise(c(7,19,21,23,50,51,85,99))
}
