% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impute_missing_values.R
\name{impute_missing_values}
\alias{impute_missing_values}
\title{Impute Missing Values}
\usage{
impute_missing_values(dataset)
}
\arguments{
\item{dataset}{A vector with missing values.}
}
\value{
A vector with missing values imputed.
}
\description{
This function imputes missing values in a vector using the mean.
}
\examples{
impute_missing_values(c(1,2,3,7,8,9,10,13,18,NA,20,NA,22,23))
}
