% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Outlier_detection.R
\name{outlier_detection}
\alias{outlier_detection}
\title{Outlier Detection}
\usage{
outlier_detection(dataset)
}
\arguments{
\item{x}{A vector.}
}
\value{
A vector indicating which values are outliers.
}
\description{
This function detects outliers in a vector using the standard IQR method.
}
\examples{
outlier_detection(c(12, 24, 36, 48, 60, 72, 84, 96, 100))
}
